style: Constant 분리
