+++ src/main/java/com/moabam/api/application/NotificationService.java
@@ -1,6 +1,7 @@
 package com.moabam.api.application;
 
-import static com.moabam.global.common.util.GlobalConstant.*;
+import static com.moabam.global.common.constant.FcmConstant.*;
+import static com.moabam.global.common.constant.GlobalConstant.*;
 import static java.util.Objects.*;
 
 import org.springframework.stereotype.Service;
