+++ src/main/java/com/moabam/global/common/constant/FcmConstant.java
@@ -1,0 +1,9 @@
+package com.moabam.global.common.constant;
+
+public class FcmConstant {
+
+	public static final String TO = "_TO_";
+	public static final long EXPIRE_KNOCK = 12;
+	public static final long EXPIRE_FCM_TOKEN = 60;
+	public static final String FIREBASE_PATH = "config/moabam-firebase.json";
+}
