+++ src/main/java/com/moabam/global/common/constant/GlobalConstant.java
@@ -1,0 +1,14 @@
+package com.moabam.global.common.constant;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class GlobalConstant {
+
+	public static final String BLANK = "";
+	public static final String COMMA = ",";
+	public static final String UNDER_BAR = "_";
+	public static final String CHARSET_UTF_8 = ";charset=UTF-8";
+	public static final String SPACE = " ";
+}
