+++ src/main/resources/static/docs/notification.html
@@ -487,7 +487,7 @@ Content-Length: 64
 <div id="footer">
 <div id="footer-text">
 Version 0.0.1-SNAPSHOT<br>
-Last updated 2023-11-04 20:22:52 +0900
+Last updated 2023-11-06 22:55:25 +0900
 </div>
 </div>
 </body>
